feat: move yeth code into separate file
